+++ cmd/root.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"errors"
 	"fmt"
+	"os"
 
 	pp "github.com/engmtcdrm/go-prettyprint"
 	"github.com/spf13/cobra"
@@ -19,6 +20,7 @@ const (
 	coverageThresholdUsage = "(optional) coverage threshold override. This will also overwrite what is specified in the configuration file"
 	debugUsage             = "(optional) enable debug output, e.g. what commands are run"
 	gitTargetRefUsage      = "(optional) git target ref to compare against (default: current branch's nearest parent branch)"
+	outputUsage            = "(optional) file to write new code missing coverage out to"
 	verboseUsage           = "(optional) enable verbose output, e.g. output from go test command"
 )
 
@@ -31,6 +33,7 @@ type cmd struct {
 	debug             bool
 	gitTargetRef      string
 	verbose           bool
+	output            string
 }
 
 func init() {
@@ -41,7 +44,7 @@ func init() {
 		Short:   app.ShortDesc,
 		Long:    app.LongDesc,
 		Example: app.Name,
-		Version: getSemVer(app.Version),
+		Version: app.Version,
 		RunE:    c.run,
 	}
 
@@ -50,8 +53,8 @@ func init() {
 	rootCmd.Flags().Float64VarP(&c.coverageThreshold, "coverage-threshold", "c", config.DefaultConfig.CoverageThreshold, coverageThresholdUsage)
 	rootCmd.Flags().BoolVarP(&c.debug, "debug", "d", false, debugUsage)
 	rootCmd.Flags().StringVarP(&c.gitTargetRef, "target-ref", "t", "", gitTargetRefUsage)
+	rootCmd.Flags().StringVarP(&c.output, "output", "o", "", outputUsage)
 	rootCmd.Flags().BoolVarP(&c.verbose, "verbose", "v", false, verboseUsage)
-
 }
 
 // Execute executes the root command.
@@ -83,22 +86,8 @@ func (c *cmd) run(cmd *cobra.Command, args []string) error {
 		fmt.Println(string(report.CoverageProfile.RawTestOutput))
 	}
 
-	if report.HasUncoveredLines() {
-		fmt.Printf("%s\n\n", colors.LightGreen("Missing coverage:"))
-
-		for _, file := range report.Files {
-			for _, lineRange := range file.UncoveredNewLineGroups {
-				fmt.Printf("%s:%s:%s\n",
-					pp.Bold(file.Path),
-					pp.Redf("%d", lineRange.Start),
-					pp.Redf("%d", lineRange.End),
-				)
-			}
-
-			if len(file.UncoveredNewLineGroups) > 0 {
-				fmt.Println()
-			}
-		}
+	if err := outputUncoveredLines(report, c.output); err != nil {
+		return err
 	}
 
 	if err != nil {
@@ -113,9 +102,12 @@ func (c *cmd) run(cmd *cobra.Command, args []string) error {
 		pp.Greenf(floatFormat, report.CoveragePercent()),
 		pp.Greenf(floatFormat, cfg.CoverageThreshold),
 	)
+
 	return nil
 }
 
+// handleFlags updates the configuration based on the command-line flags that
+// were set.
 func (c *cmd) handleFlags(cfg *config.Config, cmd *cobra.Command) {
 	if cmd.Flags().Changed("coverage-threshold") {
 		cfg.CoverageThreshold = c.coverageThreshold
@@ -127,5 +119,63 @@ func (c *cmd) handleFlags(cfg *config.Config, cmd *cobra.Command) {
 
 	if cmd.Flags().Changed("target-ref") {
 		cfg.GitDiffOptions.TargetRef = c.gitTargetRef
+	}
+}
+
+// outputUncoveredLines writes the uncovered lines from the report to
+// [os.Stdout]. If an output file path is specified, the uncovered lines will
+// also be written to that file.
+func outputUncoveredLines(report *analyzer.Report, outputFilePath string) error {
+	if !report.HasUncoveredLines() {
+		return nil
+	}
+
+	var outputFile *os.File
+
+	if outputFilePath != "" {
+		var err error
+		outputFile, err = os.Create(outputFilePath)
+		if err != nil {
+			return fmt.Errorf("failed to create output file: %w", err)
+		}
+		defer func() {
+			_ = outputFile.Close()
+		}()
+	}
+
+	fmt.Printf("%s\n\n", colors.LightGreen("Missing coverage:"))
+
+	for _, file := range report.Files {
+		for _, lineRange := range file.UncoveredNewLineGroups {
+			outputUncoveredLineToStdout(file.Path, lineRange)
+			outputUncoveredLinetoFile(outputFile, file.Path, lineRange)
+		}
+
+		if len(file.UncoveredNewLineGroups) > 0 {
+			fmt.Println()
+		}
+	}
+
+	return nil
+}
+
+// outputUncoveredLineToStdout writes the uncovered line range for a given file
+// to the [os.Stdout].
+func outputUncoveredLineToStdout(filePath string, lineRange analyzer.LineRange) {
+	fmt.Printf("%s:%s:%s\n",
+		pp.Bold(filePath),
+		pp.Redf("%d", lineRange.Start),
+		pp.Redf("%d", lineRange.End),
+	)
+}
+
+// outputUncoveredLinetoFile writes the uncovered line range for a given file to
+// the specified output file.
+func outputUncoveredLinetoFile(file *os.File, filePath string, lineRange analyzer.LineRange) {
+	switch file {
+	case nil:
+		return
+	default:
+		_, _ = fmt.Fprintf(file, "%s:%d:%d\n", filePath, lineRange.Start, lineRange.End)
 	}
 }