+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- feat: added flag `-o/--output` to write new code missing coverage out to a file
+
 ### Changed
 
 - refactor: changed header output to utilize bytes.Buffer for better performance